Refactor focus keyphrase field: replace `KeywordInput` with `TopicInputs`, update validation handling, simplify tests, and remove unused components.
